fix: long

+++ doctr/models/recognition/crnn/pytorch.py
@@ -142,7 +142,7 @@ class CRNN(RecognitionModel, nn.Module):
         logits = model_output.permute(1, 0, 2)
         probs = F.log_softmax(logits, dim=-1)
         ctc_loss_fn = nn.CTCLoss(blank=len(self.vocab))
-        ctc_loss = ctc_loss_fn(probs, torch.IntTensor(gt), input_length, torch.IntTensor(seq_len))
+        ctc_loss = ctc_loss_fn(probs, gt.to(dtype=torch.long), input_length, seq_len.to(dtype=torch.long))
 
         return ctc_loss
